Refactor callsigns.py into a callsigns directory
